pyproject.toml
[project]
name = "efund-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.6",
    "aiosignal==1.3.2",
    "akracer==0.0.13",
    "akshare==1.16.96",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "authlib==1.6.0",
    "beautifulsoup4==4.13.4",
    "brotli==1.1.0",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "comm==0.2.2",
    "cryptography==45.0.3",
    "cssselect2==0.8.0",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "deprecated==1.2.18",
    "distro==1.9.0",
    "docstring-parser==0.16",
    "et-xmlfile==2.0.0",
    "executing==2.2.0",
    "fastapi==0.115.12",
    "filelock==3.18.0",
    "fonttools==4.58.1",
    "frozenlist==1.6.0",
    "fsspec==2025.5.1",
    "google-adk==1.1.0",
    "google-api-core==2.24.2",
    "google-api-python-client==2.170.0",
    "google-auth==2.40.2",
    "google-auth-httplib2==0.2.0",
    "google-cloud-aiplatform==1.94.0",
    "google-cloud-bigquery==3.33.0",
    "google-cloud-core==2.4.3",
    "google-cloud-resource-manager==1.14.2",
    "google-cloud-secret-manager==2.23.3",
    "google-cloud-speech==2.32.0",
    "google-cloud-storage==2.19.0",
    "google-cloud-trace==1.16.1",
    "google-crc32c==1.7.1",
    "google-genai==1.17.0",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    "graphviz==0.20.3",
    "greenlet==3.2.2",
    "grpc-google-iam-v1==0.14.2",
    "grpcio==1.71.0",
    "grpcio-status==1.71.0",
    "h11==0.16.0",
    "hf-xet==1.1.2",
    "html5lib==1.1",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.32.3",
    "idna==3.10",
    "importlib-metadata==8.6.1",
    "ipykernel==6.29.5",
    "ipython==9.3.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jsonpath==0.82.2",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "litellm==1.72.0",
    "loguru==0.7.3",
    "lxml==5.4.0",
    "markdown>=3.8",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "mcp-query-table==0.3.8",
    "mcp[cli]==1.9.1",
    "mdurl==0.1.2",
    "multidict==6.4.4",
    "nest-asyncio==1.6.0",
    "notebook>=7.4.3",
    "numpy==2.2.6",
    "openai==1.83.0",
    "openpyxl==3.1.5",
    "opentelemetry-api==1.33.1",
    "opentelemetry-exporter-gcp-trace==1.9.0",
    "opentelemetry-resourcedetector-gcp==1.9.0a0",
    "opentelemetry-sdk==1.33.1",
    "opentelemetry-semantic-conventions==0.54b1",
    "packaging==25.0",
    "pandas==2.2.3",
    "pandoc==2.4",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.2.1",
    "platformdirs==4.3.8",
    "playwright==1.52.0",
    "playwright-stealth==1.0.6",
    "plumbum==1.9.0",
    "ply==3.11",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.1",
    "proto-plus==1.26.1",
    "protobuf==5.29.4",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "py-mini-racer==0.6.0",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pydyf==0.11.0",
    "pyee==13.0.0",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "pyphen==0.17.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==26.4.0",
    "referencing==0.36.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rich==14.0.0",
    "rpds-py==0.25.1",
    "rsa==4.9.1",
    "schedule==1.2.2",
    "setuptools==80.9.0",
    "shapely==2.1.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.41",
    "sse-starlette==2.3.5",
    "stack-data==0.6.3",
    "starlette==0.46.2",
    "tabulate==0.9.0",
    "tiktoken==0.9.0",
    "tinycss2==1.4.0",
    "tinyhtml5==2.0.0",
    "tokenizers==0.21.1",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typer==0.16.0",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "uritemplate==4.1.1",
    "urllib3==2.4.0",
    "uvicorn==0.34.2",
    "wcwidth==0.2.13",
    "weasyprint==65.1",
    "webencodings==0.5.1",
    "websockets==15.0.1",
    "wrapt==1.17.2",
    "xlrd==2.0.1",
    "yarl==1.20.0",
    "zipp==3.22.0",
    "zopfli==0.2.3.post1",
]
